Fixed compilation issues when using more recent versions of mingw
An old and hackish 'usleep' define in 'stdafx.h' started to clash with
mingw headers for the recent versions of mingw (because now mingw
provides 'usleep' function itself). The solution was to just switch
to the use of allegro-specific 'rest' function which provides the same
functionality.

It's a temporary solution anyway, because any occurences of
usleep/rest functions should be eliminated completely in order to
reduce power consumption. Low power consumption is important
for battery powered devices (laptops and smartphones).

Author: Siarhei Siamashka <siarhei.siamashka@gmail.com>