Makefile: do not bail out if 'svn delete --force ...' returns error
This is a part of different make targets (building of win32 installer
for example). If the directory to be deleted did not exist, svn
returns error now (it did not seem to do this in older versions)
and make stops. just ignoring this error helps.

From: Siarhei Siamashka <siarhei.siamashka@gmail.com>