Disable FULLSCREEN and SCALE2X in the default configuration
Nowadays widescreen monitors are quite common and the
game may have wrong aspect ratio in fullscreen mode. Also
the use of scale2x is questionable because it distorts the
graphics.

Anyway, this seems to be a bit more sane default. Both
fullscreen and scale2x modes are tunable from the game
itself. The rest of settings are in ufo200.ini file which is
also editable.

From: Siarhei Siamashka <siarhei.siamashka@gmail.com>